Déplacement des tests dans le bon package
Les tests maven ne se placent pas au même endroit qu'en java classique.
https://maven.apache.org/guides/introduction/introduction-to-the-standard-directory-layout.html
